+++ flutter_google_places_sdk/lib/flutter_google_places_sdk.dart
@@ -7,8 +7,9 @@ export 'package:flutter_google_places_sdk_platform_interface/flutter_google_plac
 /// Client used to call methods on the native google places sdk
 class FlutterGooglePlacesSdk {
   /// Construct a FlutterGooglePlacesSdk using the specific api key and locale
-  FlutterGooglePlacesSdk(this._apiKey, {Locale? locale})
-      : this._locale = locale;
+  FlutterGooglePlacesSdk(this._apiKey, {Locale? locale, bool? useNewApi})
+      : this._locale = locale,
+        this._useNewApi = useNewApi;
 
   /// "Powered by google" image that should be used when background is white
   static const AssetImage ASSET_POWERED_BY_GOOGLE_ON_WHITE =
@@ -19,8 +20,7 @@ class FlutterGooglePlacesSdk {
       FlutterGooglePlacesSdkPlatform.ASSET_POWERED_BY_GOOGLE_ON_NON_WHITE;
 
   /// Singleton instance to the platform
-  static FlutterGooglePlacesSdkPlatform platform =
-      FlutterGooglePlacesSdkPlatform.instance;
+  static FlutterGooglePlacesSdkPlatform platform = FlutterGooglePlacesSdkPlatform.instance;
 
   /// The Places API key
   /// https://developers.google.com/maps/documentation/places/android-sdk/get-api-key
@@ -32,6 +32,11 @@ class FlutterGooglePlacesSdk {
   Locale? get locale => _locale;
 
   Locale? _locale;
+
+  /// Whether the Places API (new) is used in the SDK
+  bool? get useNewApi => _useNewApi;
+
+  bool? _useNewApi;
 
   Future<void>? _lastMethodCall;
   Future<void>? _initialization;
@@ -65,7 +70,7 @@ class FlutterGooglePlacesSdk {
   }
 
   Future<void> _ensureInitialized() {
-    return _initialization ??= platform.initialize(apiKey, locale: locale)
+    return _initialization ??= platform.initialize(apiKey, locale: locale, useNewApi: useNewApi)
       ..catchError((dynamic err) {
         print('FlutterGooglePlacesSdk::_ensureInitialized error: $err');
         _initialization = null;
@@ -74,7 +79,7 @@ class FlutterGooglePlacesSdk {
 
   /// Fetches autocomplete predictions based on a query.
   ///
-  /// If [countries]] is not null, the results are restricted to those countries.
+  /// If [countries] is not null, the results are restricted to those countries.
   ///
   /// [AutocompletePrediction.distanceMeters] is calculated from [origin],
   /// if available. If [origin] is null, no distances will be included in the
@@ -115,8 +120,7 @@ class FlutterGooglePlacesSdk {
   /// Note that different fields can incur different billing.
   ///
   /// For more info about billing: https://developers.google.com/maps/documentation/places/web-service/usage-and-billing
-  Future<FetchPlaceResponse> fetchPlace(String placeId,
-      {required List<PlaceField> fields}) {
+  Future<FetchPlaceResponse> fetchPlace(String placeId, {required List<PlaceField> fields}) {
     return _addMethodCall(() => platform.fetchPlace(placeId, fields: fields));
   }
 
@@ -126,10 +130,8 @@ class FlutterGooglePlacesSdk {
   /// together with the [PlaceField.PhotoMetadatas] field
   ///
   /// For more info: https://developers.google.com/maps/documentation/places/android-sdk/photos
-  Future<FetchPlacePhotoResponse> fetchPlacePhoto(PhotoMetadata photoMetadata,
-      {int? maxWidth, int? maxHeight}) {
-    return _addMethodCall(() => platform.fetchPlacePhoto(photoMetadata,
-        maxWidth: maxWidth, maxHeight: maxHeight));
+  Future<FetchPlacePhotoResponse> fetchPlacePhoto(PhotoMetadata photoMetadata, {int? maxWidth, int? maxHeight}) {
+    return _addMethodCall(() => platform.fetchPlacePhoto(photoMetadata, maxWidth: maxWidth, maxHeight: maxHeight));
   }
 
   /// Returns whether or not the client has been initialized.
@@ -140,11 +142,83 @@ class FlutterGooglePlacesSdk {
   /// Updates the settings of the places client with the given API key and locale.
   /// If apiKey is null, the last key will be used.
   /// If locale is null, it will not be updated.
-  Future<void> updateSettings({String? apiKey, Locale? locale}) {
+  Future<void> updateSettings({String? apiKey, Locale? locale, bool? useNewApi}) {
     _apiKey = apiKey ?? this.apiKey;
     _locale = locale;
+    _useNewApi = useNewApi ?? this.useNewApi;
 
-    return _addMethodCall(
-        () => platform.updateSettings(_apiKey, locale: locale));
+    return _addMethodCall(() => platform.updateSettings(_apiKey, locale: locale, useNewApi: useNewApi));
+  }
+
+  /// Fetches places based on an ambiguous text query.
+  ///
+  /// Only the requested [fields] will be returned. If none specified,
+  /// all fields will be returned.
+  ///
+  /// Note that different fields can incur different billing.
+  ///
+  /// For more info about billing: https://developers.google.com/maps/documentation/places/android-sdk/usage-and-billing#pricing-new
+  ///
+  /// For more info on text search: https://developers.google.com/maps/documentation/places/android-sdk/text-search
+  Future<SearchByTextResponse> searchByText(
+    String textQuery, {
+    required List<PlaceField> fields,
+    String? includedType,
+    int? maxResultCount,
+    LatLngBounds? locationBias,
+    LatLngBounds? locationRestriction,
+    double? minRating,
+    bool? openNow,
+    List<int>? priceLevels,
+    TextSearchRankPreference? rankPreference,
+    String? regionCode,
+    bool? strictTypeFiltering,
+  }) {
+    return _addMethodCall(() => platform.searchByText(
+          textQuery,
+          fields: fields,
+          includedType: includedType,
+          maxResultCount: maxResultCount,
+          locationBias: locationBias,
+          locationRestriction: locationRestriction,
+          minRating: minRating,
+          openNow: openNow,
+          priceLevels: priceLevels,
+          rankPreference: rankPreference,
+          regionCode: regionCode,
+          strictTypeFiltering: strictTypeFiltering,
+        ));
+  }
+
+  /// Search for place(s) of interest using a location.
+  ///
+  /// Only the requested [fields] will be returned. If none specified,
+  /// all fields will be returned.
+  ///
+  /// Note that different fields can incur different billing.
+  ///
+  /// For more info on nearby search: https://developers.google.com/maps/documentation/places/android-sdk/nearby-search
+  Future<SearchNearbyResponse> searchNearby({
+    required List<PlaceField> fields,
+    required CircularBounds locationRestriction,
+    List<String>? includedTypes,
+    List<String>? includedPrimaryTypes,
+    List<String>? excludedTypes,
+    List<String>? excludedPrimaryTypes,
+    NearbySearchRankPreference? rankPreference,
+    String? regionCode,
+    int? maxResultCount,
+  }) {
+    return _addMethodCall(() => platform.searchNearby(
+          fields: fields,
+          locationRestriction: locationRestriction,
+          includedTypes: includedTypes,
+          includedPrimaryTypes: includedPrimaryTypes,
+          excludedTypes: excludedTypes,
+          excludedPrimaryTypes: excludedPrimaryTypes,
+          rankPreference: rankPreference,
+          regionCode: regionCode,
+          maxResultCount: maxResultCount,
+        ));
   }
 }