+++ flutter_google_places_sdk_http/lib/flutter_google_places_sdk_http.dart
@@ -20,6 +20,7 @@ class FlutterGooglePlacesSdkHttpPlugin
 
   String? _apiKey;
   Locale? _locale;
+  bool? _useNewApi;
 
   String? _lastSessionToken;
 
@@ -27,22 +28,27 @@ class FlutterGooglePlacesSdkHttpPlugin
   Future<void> deinitialize() async {
     _apiKey = null;
     _locale = null;
+    _useNewApi = null;
   }
 
   @override
-  Future<void> initialize(String apiKey, {Locale? locale}) async {
+  Future<void> initialize(String apiKey, {Locale? locale, bool? useNewApi}) async {
     _apiKey = apiKey;
     _locale = locale;
+    _useNewApi = useNewApi;
   }
 
   @override
   Future<bool?> isInitialized() async => _apiKey != null;
 
   @override
-  Future<void> updateSettings(String apiKey, {Locale? locale}) async {
+  Future<void> updateSettings(String apiKey, {Locale? locale, bool? useNewApi}) async {
     _apiKey = apiKey;
     if (locale != null) {
       _locale = locale;
+    }
+    if (useNewApi != null) {
+      _useNewApi = useNewApi;
     }
   }